fix exterior BOTI doors not showing alarm color when unpowered
When an exterior door is open during an alarm *and* there is no power,
then the windows will be black, whereas all other windows are dark red.

This commit fixes it so that an exterior open door will show with the
correct (dark red) color in the BOTI when alarms are on while the
TARDIS is not powered.