fix TARDIS sometimes landing in floor when summoned
When landing type is disabled, then pointing the sonic (in TARDIS mode)
at the floor beneath you (or at half a block away from the one you are
standing on) will make the TARDIS land one block in the floor.

This commit changes the way the Y coordinate is determined, so that
this no longer happens.