fix GUI buttons
The buttons have a lot of deadzones where clicking the button does not
work. The clickable area is also sometimes going out of the button's
bounds or partially overlapping with other buttons.
Also, the interior settings button also gets highlighted together with
some other buttons.
This commit fixes all that.